Add adjust ui

+++ Utiltity/time_utility.py
@@ -118,7 +118,7 @@ class Clock:
     def elapsed_s(self) -> int:
         return int(time.time() - self.__start)
 
-    def elapsed_ms(self) -> float:
+    def elapsed_ms(self) -> int:
         return int((time.time() - self.__start) * 1000)
